+++ src/services/BlockchainScanner.ts
@@ -19,12 +19,18 @@ export abstract class BlockchainScanner {
   blockchainId!: string;
   markerId!: string;
   step!: number;
-  provider!: StaticJsonRpcProvider;
+  provider: StaticJsonRpcProvider | undefined;
 
   abstract apply(fromBlock: number, toBlock: number): Promise<boolean>;
 
   async run(): Promise<Result> {
     const fromBlock = await this.getMarker();
+
+    if (!this.provider) {
+      this.logger.info(`[BlockchainScanner] not active`);
+      return {success: false, fromBlock};
+    }
+
     const toBlock = await this.getToBlock(fromBlock);
     let success = false;
 
@@ -39,6 +45,8 @@ export abstract class BlockchainScanner {
   }
 
   private async getToBlock(fromBlock: number): Promise<number> {
+    if (!this.provider) throw new Error(`[BlockchainScanner] getToBlock: no provider`);
+
     const latestBlock = await this.provider.getBlockNumber();
     const stepBlock = fromBlock + this.step;
     return latestBlock > stepBlock ? stepBlock : latestBlock;